Conflicts:
	DataFormats/TrackerRecHit2D/interface/SiTrackerGSMatchedRecHit2D.h
	DataFormats/TrackerRecHit2D/src/classes_def.xml

+++ Mixing/Base/interface/PileUp.h
@@ -149,7 +149,7 @@ namespace edm {
     RecordEventID(std::vector<edm::SecondaryEventIDAndFileInfo>& ids, T& eventOperator)
       : ids_(ids), eventOperator_(eventOperator), eventCount(0) {
     }
-    void operator()(EventPrincipal const& eventPrincipal, unsigned int fileNameHash) {
+    void operator()(EventPrincipal const& eventPrincipal, size_t fileNameHash) {
       ids_.emplace_back(eventPrincipal.id(), fileNameHash);
       eventOperator_(eventPrincipal, ++eventCount);
     }
